Issue #21 - Refactor speedometer backend APIs
We're refactoring the backend to use query parameters and use modeIds
to filter through the data returned by AWFY. This guarantees that if
the order of an array changes on AWFY we won't be grabbing data based on
the index of an array.

Instead of having multiples speedometer APIs you now call the speedometer API with
few parameters. For example:
http://localhost:3000/api/perf/benchmark/speedometer?channel=nightly&architecture=64

The data returned now has metadata regarding how it was generated. For example:
```json
  "meta": {
    "configuration": {
      "Firefox": 14,
      "Chrome": 3
    },
    "dataUrl": "https://arewefastyet.com/data.php?file=bk-aggregate-speedometer-misc-score-36.json",
    "viewUrl": "https://arewefastyet.com#machine=36&view=single&suite=speedometer-misc&subtest=score"
  },
```