Make Hot Module Replacement actually work
I had noticed that changes on the API would not be reflected when reloading
the endpoints.

If I understand correctly, when HMR notices the module change it accepts it,
however, it needs to close the previous server and create a new one.
The old server had a reference to the old `app` module and then a new
server is created referencing the newly loaded instance.