Refactor Speedometer API to accept many more options
A lot of the code added in this change comes from the frontend code.
Instead of doing all sorts of manipulations on the frontend we can
instead rely on the backend to give us the correct data.

This change also allows to simply raise an error and the API will
return the correct message to the user.