Script to fetch Nimbledroid data
This script will cache all Nimbledroid data for focus and klar in Redis.

The two fetches are only cached for 30 minutes since the number of code
landing for focus/klar is approximately that. If Nimbledroid's profiling runs
throughput at the same speed we should get the most recent data.

In the future, the API will have pagination (about 10 runs) and will give us
the number of the most recent build. The code in this patch is already taken
some of that into account and will make switching to it require less development
work.

Only profiling runs that are completed (marked as "Failed") are cached indefinitively.